Add IdRequestInterface and IdRequestTrait
in order to improve behaviour of replacing id in urls starting from PHP 8
