+++ lib/Wsc/Request/LiveStreamState/ResetRequest.php
@@ -2,33 +2,19 @@
 
 namespace Ticketpark\Wsc\Request\LiveStreamState;
 
+use Ticketpark\Wsc\Request\IdRequestInterface;
+use Ticketpark\Wsc\Request\IdRequestTrait;
 use Ticketpark\Wsc\Request\Request;
 use Ticketpark\Wsc\Request\RequestCommonsTrait;
 use Ticketpark\Wsc\Response\LiveStreamState\FetchResponse;
 
-class ResetRequest extends Request
+class ResetRequest extends Request implements IdRequestInterface
 {
     const API_PATH = '/live_streams/{id}/reset';
     const RESPONSE_CLASS = FetchResponse::class;
 
     use RequestCommonsTrait;
-
-    /**
-     * @var string
-     */
-    private $id;
-
-    public function getId(): ?string
-    {
-        return $this->id;
-    }
-
-    public function setId(string $id): self
-    {
-        $this->id = $id;
-
-        return $this;
-    }
+    use IdRequestTrait;
 
     public function getHttpMethod(): string
     {
